% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{adjust_mean_variance}
\alias{adjust_mean_variance}
\title{Adjust a variable to have equal mean and variance across all levels of another variable}
\usage{
adjust_mean_variance(y, x, keep.scale = TRUE)
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Independent variable}

\item{keep.scale}{Default = TRUE. Keep the original mean and sd}
}
\value{
Variable
}
\description{
Adjust a variable to have equal mean and variance across all levels of another variable
}
