% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_from_pyscript.R
\name{att_from_pyscript}
\alias{att_from_pyscript}
\title{Extract imports from Python script}
\usage{
att_from_pyscript(file)
}
\arguments{
\item{file}{File path to a .py file}
}
\value{
A comma-separated character string.
}
\description{
This function is inspired by \code{att_from_rscript()} in the attachment
package. It assumes that Python imports start on line 1, are not separated
by any empty lines, and that an empty line exists between imports and code.
}
