% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_params.R
\name{beta_params}
\alias{beta_params}
\title{Beta Distribution Parameters}
\usage{
beta_params(mu, var)
}
\arguments{
\item{mu}{A numeric value specifying the mean.}

\item{var}{A numeric value specifying the variance.}
}
\value{
A list of length two containing the two shape parameters.
}
\description{
This function computes shape parameters of a beta distribution, given the mean and variance.
}
