% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_palette.R
\name{blend_palette}
\alias{blend_palette}
\title{Blend a colour palette with another colour}
\usage{
blend_palette(palette, blend_colour, view_palette = FALSE)
}
\arguments{
\item{palette}{The starting vector for the palette, which must be either be
recognised colour names (e.g. "blue") or hex colour codes (e.g. "#ffffff").}

\item{blend_colour}{The colour to blend with, which must be either be a
recognised colour name (e.g. "blue") or a hex colour code (e.g. "#ffffff").}

\item{view_palette}{Boolean whether or not to view the palette.}
}
\value{
A character vector of colours.
}
\description{
This function allows users to blend a colour palette vector with another
colour to make a more cohesive palette.
}
\examples{
blend_palette(c("red", "purple"), "blue")
}
