% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{check_rootname}
\alias{check_rootname}
\title{Check if files exist}
\usage{
check_rootname(rootname, software = "plink", suffix = "all")
}
\arguments{
\item{rootname}{rootname}

\item{software}{Either "plink" or "gcta"}

\item{suffix}{Either "all" or a specific suffic}
}
\value{
Character string
}
\description{
Either checks of all files for a rootname exist, or returns a specific suffix
}
