% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{comma}
\alias{comma}
\title{Format numbers}
\usage{
comma(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
character vector with all numbers formated
}
\description{
Format numbers to being 2 digits
}
