% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffs.R
\name{diffs}
\alias{diffs}
\title{First Differences}
\usage{
diffs(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
This function is the opposite of cumsum which returns the
first order differences between vector elements. An extension
of diff, also including the first element of the vector.
}
\examples{
diffs(c(1,5,7))
}
