% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{err_msg}
\alias{err_msg}
\title{Function for dealing with error messages when using tryCatch}
\usage{
err_msg(e, r_msg = TRUE, user_msg = NULL, to_return = NA)
}
\arguments{
\item{e}{error message from function}

\item{r_msg}{print the error message given by the function}

\item{user_msg}{a message given by the user}

\item{to_return}{what should be returned if there is an error}
}
\value{
what is chosen by the user to be returned, default = NA
}
\description{
Function for dealing with error messages when using tryCatch
}
