% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac-addition.R
\name{extract_alspac_labels}
\alias{extract_alspac_labels}
\title{extract_alspac_labels}
\usage{
extract_alspac_labels(dat, alsp_dir)
}
\arguments{
\item{dat}{data.frame with columns "path", "obj", name", that correspond to the path within the ALSPAC directory to find the files, the file names of interest and the variables of interest.}

\item{alsp_dir}{string. Path to ALSPAC directory on local machine.}
}
\value{
list of variable names and their labels
}
\description{
Extract Stata labels from ALSPAC .dta files. These are needed to interpret numeric values assigned to categorical variables.
}
