% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{extract_sum_stats}
\alias{extract_sum_stats}
\title{Extract summary statistics from numeric vector}
\usage{
extract_sum_stats(
  dat,
  min = T,
  q1 = T,
  med = T,
  mean = T,
  q3 = T,
  max = T,
  na = T
)
}
\arguments{
\item{dat}{numeric vector}

\item{min}{extract minimum value}

\item{q1}{extract first quartile}

\item{med}{extract median}

\item{q3}{extract third quartile}

\item{max}{extract maximum value}

\item{na}{extract number of NA values}
}
\value{
data.frame of summary stats selected
}
\description{
Extract summary statistics from numeric vector
}
