% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forest_plot}
\alias{forest_plot}
\title{Generate a forest plot}
\usage{
forest_plot(
  dat,
  col_num,
  group = NA,
  y_axis,
  units = NULL,
  title = NULL,
  scale = 1,
  null_at = 1,
  text_size = "norm",
  meta = FALSE,
  f_var = "facet_var"
)
}
\arguments{
\item{dat}{a data.frame containing the data}

\item{group}{group}

\item{y_axis}{y-axis}

\item{units}{units}

\item{title}{title}

\item{scale}{scale}

\item{null_at}{null at}

\item{text_size}{text size}

\item{meta}{meta-analysis?}

\item{f_var}{facet variable name}
}
\value{
forest plot
}
\description{
Generate a forest plot using ggplot2. This function has capability to be split into multiple forests and was generated to display Mendelian randomization results with a binary outcome.
}
