% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_docs_table.R
\name{function_docs_table}
\alias{function_docs_table}
\title{Find title and description of all exported objects in package documentation}
\usage{
function_docs_table(pkg_name)
}
\arguments{
\item{pkg_name}{Package name given as a string}
}
\value{
Tibble with three columns
}
\description{
Find title and description of all exported objects in package documentation
}
