% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geom_vio_box}
\alias{geom_vio_box}
\title{Generate a boxplot within a violin plot}
\usage{
geom_vio_box(p, width = 0.2, ...)
}
\arguments{
\item{p}{a plot with the class "ggplot"}

\item{width}{width of the boxplot within the violin plot. Default = 0.2}

\item{...}{other parameters that can be parsed to geom_violin and geom_boxplot}
}
\value{
boxplot within a violin plot
}
\description{
Generate a boxplot within a violin plot
}
