% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_cb_palette}
\alias{get_cb_palette}
\title{Get colours for plotting that are better for colour blind people}
\usage{
get_cb_palette()
}
\value{
a vector of colour IDs
}
\description{
Generates a vector of colour IDs that can be used for plots
}
