% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_docs_table.R
\name{get_docs}
\alias{get_docs}
\title{Find title and description of function in package documentation}
\usage{
get_docs(function_name, pkg_name)
}
\arguments{
\item{function_name}{Function name given as a string}

\item{pkg_name}{Package name given as a string}
}
\value{
Tibble with three columns
}
\description{
Find title and description of function in package documentation
}
