% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{get_residuals}
\alias{get_residuals}
\title{Get residuals and keep original scale}
\usage{
get_residuals(formula, data, keep.scale = TRUE, ...)
}
\arguments{
\item{formula}{Formula}

\item{data}{Data frame}

\item{keep.scale}{Keep original mean and sd of y variable. Default = TRUE}

\item{...}{Extra variables to pass to lm}
}
\value{
Variable adjusted for independent variables
}
\description{
Get residuals and keep original scale
}
