% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{getmode}
\alias{getmode}
\title{Extract mode from a numeric vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
mode of numeric vector
}
\description{
Extract mode from a numeric vector
}
