% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector of elements to check}

\item{y}{vector to check against}
}
\value{
A boolean vector.
}
\description{
This function negates the in function and returns true if
the item before notin is not in the item after.
}
