% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_to_decimal.R
\name{hms_to_decimal}
\alias{hms_to_decimal}
\title{Hours:Minutes:Seconds to Decimal}
\usage{
hms_to_decimal(data)
}
\arguments{
\item{data}{Input vector or element in hh:mm:ss format.}
}
\value{
A numeric vector of times in hours.
}
\description{
This function converts a time object in the hh:mm:ss form to a numeric in hours.
}
