% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{impute_matrix}
\alias{impute_matrix}
\title{Impute missing values in DNAm matrix}
\usage{
impute_matrix(x, FUN = function(x) matrixStats::rowMedians(x, na.rm = T))
}
\arguments{
\item{x}{DNAm matrix}

\item{FUN}{function to apply to "x" to get values to impute missing values}
}
\value{
imputed DNAm matrix
}
\description{
Impute missing values in DNAm matrix
}
