% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{is.binary}
\alias{is.binary}
\title{Check if a vector is binary}
\usage{
is.binary(v)
}
\arguments{
\item{v}{vector}
}
\value{
TRUE or FALSE depending on whether the vector is binary
}
\description{
Check if a vector is binary
}
