% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{is.monomorphic}
\alias{is.monomorphic}
\title{Check if a vector is monomorphic}
\usage{
is.monomorphic(v)
}
\arguments{
\item{v}{vector}
}
\value{
TRUE or FALSE depending on whether the vector is monomorphic
}
\description{
Check if a vector is monomorphic
}
