% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{last}
\usage{
last(x, type = "col")
}
\arguments{
\item{x}{Vector or dataframe to extract last element from}

\item{type}{Character specifying whether to use last row or column if
dataframe. One of "row" or "col". Default "col".}
}
\value{
A vector or dataframe.
}
\description{
Function to extract last element
}
\examples{
last(data.frame(x = c(1, 2, 3), y = c("A", "B", "C")))
}
