% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenotypes.R
\name{make_geno}
\alias{make_geno}
\title{Create a genotype or many genotypes in HWE}
\usage{
make_geno(n, p)
}
\arguments{
\item{n}{Sample size}

\item{p}{Array of allele frequencies to simulate. Can be a single value.}
}
\value{
Array or matrix of genotypes
}
\description{
Under Hardy Weinberg a genotype is binomially distributed, with p = allele frequency
}
\examples{
\dontrun{
g <- make_geno(100, 0.5) # To make a single genotype
g <- make_geno(100, c(0.5, 0.3, 0.6)) # To make 3 genotypes in a matrix
}
}
