% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{make_grm_matrix}
\alias{make_grm_matrix}
\title{Convert long GRM format to matrix format}
\usage{
make_grm_matrix(grm)
}
\arguments{
\item{grm}{Result from readGRM}
}
\value{
Matrix of n x n
}
\description{
Convert long GRM format to matrix format
}
