% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{make_pretty}
\alias{make_pretty}
\title{Format a set of numbers to have a certain number of sig figs}
\usage{
make_pretty(num, digits)
}
\arguments{
\item{num}{numeric vector}

\item{digits}{number of digits}
}
\value{
character vector with all numbers formated
}
\description{
Format a set of numbers to have a certain number of sig figs
}
