% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{make_pretty_lm_table}
\alias{make_pretty_lm_table}
\title{Format a set of numbers to have a certain number of sig figs}
\usage{
make_pretty_lm_table(
  dat,
  voi,
  est = "Estimate",
  se = "StdErr",
  upper_CI = "97.5 \%",
  lower_CI = "2.5 \%",
  p = "Pr(>|t|)",
  arrange_by = p,
  output = c("se", "CI")
)
}
\arguments{
\item{dat}{data}

\item{voi}{variable of interest}

\item{est}{effect estimate variable}

\item{se}{SE variable}

\item{upper_CI}{upper confidence interval variable name}

\item{lower_CI}{lower confidence interval variable name}

\item{p}{pvalue variable name}

\item{arrange_by}{variable to arrange output by}

\item{output}{either SE or CI or both in output}
}
\value{
character vector with all numbers formated
}
\description{
Format a set of numbers to have a certain number of sig figs
}
