% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{new_load}
\alias{new_load}
\title{Use the base R load function but name the object(s) being loaded in}
\usage{
new_load(file)
}
\arguments{
\item{file}{file name}
}
\value{
data loaded in
}
\description{
Use the base R load function but name the object(s) being loaded in
}
