% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_report.R
\name{num_to_text}
\alias{num_to_text}
\title{Convert numbers to text}
\usage{
num_to_text(x, start_of_sentence = FALSE)
}
\arguments{
\item{x}{a single integer to have converted into text if needs be}

\item{start_of_sentence}{TRUE or FALSE to indicate if the number is at the start of a sentence. Default = FALSE}
}
\value{
Number written in text or number format
}
\description{
Convert numbers to text when needed in rmarkdown reports
}
