% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{plot_theme}
\usage{
plot_theme(main_font = "sans", title_font = "sans", hjust = 0)
}
\arguments{
\item{main_font}{Font used for all text elements except the title. Default "sans".}

\item{title_font}{Font used for the title. Default "sans".}

\item{hjust}{Horizontal alignment of title, subtitle, and caption. Default 0.}
}
\value{
A ggplot2 theme
}
\description{
Custom ggplot theme
}
