% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{pool_mean_sd}
\alias{pool_mean_sd}
\title{Pool mean and SD estimates from two groups or cohorts}
\usage{
pool_mean_sd(n1, n2, m1, m2, s1, s2)
}
\arguments{
\item{n1}{sample size of group 1}

\item{n2}{sample size of group 2}

\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{s1}{SD of group 1}

\item{s2}{SD of group 2}
}
\value{
list with pooled mean and SD estimates
}
\description{
Pool mean and SD estimates from two groups or cohorts
}
