% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_hex.R
\name{random_hex}
\alias{random_hex}
\title{Generates a random hex code}
\usage{
random_hex(n)
}
\arguments{
\item{n}{Number of hex codes to generate}
}
\value{
Character string of hex codes
}
\description{
Generates a random hex code
}
\examples{
random_hex(2)
}
