% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{rank_transform}
\alias{rank_transform}
\title{Rank transform variable}
\usage{
rank_transform(x, s = NULL, m = NULL)
}
\arguments{
\item{x}{Numeric variable}

\item{s}{Outcome standard deviation. Optional}

\item{m}{Outcome mean. Optional}
}
\value{
Variable
}
\description{
Convert variable into theoretical normal quantiles
}
