% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{read_delim_rownames}
\alias{read_delim_rownames}
\title{Read in files with rownames using read_delim() from the readr package}
\usage{
read_delim_rownames(input, delim, ...)
}
\arguments{
\item{input}{file name}

\item{delim}{delimiter}

\item{...}{any parameters for read_delim()}
}
\value{
tibble of data that has rownames
}
\description{
Read in files with rownames using read_delim() from the readr package
}
