% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{read_gcta_xmat}
\alias{read_gcta_xmat}
\title{Read GCTA xmat file}
\usage{
read_gcta_xmat(filename)
}
\arguments{
\item{filename}{}
}
\value{
List with the following elements:
        xmat: Matrix of nsnp x nid 0/1/2
        snps: SNP IDs (array)
        ids: FID and IID
}
\description{
Read GCTA xmat file
}
