% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{read_grm}
\alias{read_grm}
\title{Read binary GCTA GRM}
\usage{
read_grm(rootname, n.file = FALSE)
}
\arguments{
\item{rootname}{Name of files. e.g. if the GRM files are data.grm.bin, data.grm.id and data.grm.N.bin then supply "data" as the argument}

\item{n.file}{Write out grm.N.bin file - not always needed}
}
\value{
List with elements:
        grm: data frame of GRM in long format
        id: data frame of IDs
}
\description{
<full description>
}
