% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{read_hsq}
\alias{read_hsq}
\title{Read in GCTA .hsq file in long format}
\usage{
read_hsq(filename, label = NULL)
}
\arguments{
\item{filename}{.hsq file}

\item{label}{Optional value for labelling the output}
}
\value{
Data frame of REML results
}
\description{
Read in GCTA .hsq file in long format
}
