% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{read_plink_raw}
\alias{read_plink_raw}
\title{Read plink raw format as exported from PLINK2 using --recode A}
\usage{
read_plink_raw(filename)
}
\arguments{
\item{filename}{Filename of exported data}
}
\value{
List with the following elements:
        xmat: Matrix of nsnp x nid 0/1/2
        snps: SNP information (rsid and ref allele)
        ids: id information (ped file)
}
\description{
Read plink raw format as exported from PLINK2 using --recode A
}
