% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{remove_outliers_iterative}
\alias{remove_outliers_iterative}
\title{Iteratively remove outliers}
\usage{
remove_outliers_iterative(x, niter = 5, d = 3)
}
\arguments{
\item{x}{Variable}

\item{niter}{Number of iterations}

\item{d}{Number of standard deviations from mean}
}
\value{
Variable
}
\description{
Remove outliers above or below d standard deviations. Repeat for niter iterations
}
