% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_df.R
\name{rep_df}
\alias{rep_df}
\title{rep_df}
\usage{
rep_df(df, times = 1, length.out = NA, each = 1)
}
\arguments{
\item{df}{The dataframe to be replicated.}

\item{times}{an integer-valued vector giving the (non-negative) number of times
to repeat each element if of length length(x), or to repeat the whole vector if
of length 1. Negative or NA values are an error. A double vector is accepted,
other inputs being coerced to an integer or double vector.}

\item{length.out}{non-negative integer. The desired length of the output vector.
Other inputs will be coerced to a double vector and the first element taken.
Ignored if NA or invalid.}

\item{each}{non-negative integer. Each element of x is repeated each times.
Other inputs will be coerced to an integer or double vector and the first element
taken. Treated as 1 if NA or invalid.}
}
\value{
A dataframe.
}
\description{
Function to replicate dataframes.
}
