% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose.}

\item{...}{inherited arguments}
}
\value{
a vector of length size with elements drawn from x
}
\description{
Resamples x of a specific size
}
