% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_google.R
\name{search_google}
\alias{search_google}
\title{Search Google}
\usage{
search_google(search_terms)
}
\arguments{
\item{search_terms}{character vector of terms to search}
}
\description{
Function to open a Google search in a browser for a set of terms
}
\examples{
search_google("twitter #rstats")
search_google(c("twitter", "#rstats"))
}
