% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_wrap_break.R
\name{str_wrap_break}
\alias{str_wrap_break}
\title{str_wrap_break}
\usage{
str_wrap_break(x, break_limit)
}
\arguments{
\item{x}{The character string (or character vector) to wrap}

\item{break_limit}{The character limit per line}
}
\value{
A character string (or vector)
}
\description{
Vectorised function to wrap strings and preserve line breaks
}
