% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{summarise_glm}
\alias{summarise_glm}
\title{Extract summary stats from glm() when doing logistic regression}
\usage{
summarise_glm(fit, outcome, exposure)
}
\arguments{
\item{fit}{regression output from glm() function}

\item{outcome}{the outcome variable}

\item{exposure}{the exposure variable}
}
\value{
data.frame containing outcomes, estimate, se, p and CIs, residuals and the input
}
\description{
Extract summary stats from glm() when doing logistic regression
}
