% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{summarise_lm}
\alias{summarise_lm}
\title{Extract summary stats from lm()}
\usage{
summarise_lm(fit, outcome, exposure)
}
\arguments{
\item{fit}{regression output from lm() function}

\item{outcome}{the outcome variable}

\item{exposure}{the exposure variable}
}
\value{
data.frame containing outcomes, estimate, se, p and CIs, residuals and the input
}
\description{
Extract summary stats from lm()
}
