% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_report.R
\name{tidy_colnames}
\alias{tidy_colnames}
\title{Tidy column names}
\usage{
tidy_colnames(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
character vector with all numbers formated
}
\description{
Swap underscores for dashes in column names of tables for rmarkdown reports
}
