% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_report.R
\name{tidy_nums}
\alias{tidy_nums}
\title{Tidy numbers}
\usage{
tidy_nums(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame with all numbers formated based on the comma function
}
\description{
Tidy numbers in a table for a rmarkdown report
}
