% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_abb_to_name.R
\name{US_abb_to_name}
\alias{US_abb_to_name}
\title{US State Abbreviation to Name}
\usage{
US_abb_to_name(x)
}
\arguments{
\item{x}{A character element or vector.}
}
\value{
A character element or vector.
}
\description{
This function converts a two-letter US state abbreviation to the state name.
It uses the base R state.abb and state.name vectors, but is extended to included
the District of Columbia.
}
\examples{
US_abb_to_name(c("TX", "DC"))
}
\seealso{
\code{\link[=US_name_to_abb]{US_name_to_abb()}}
}
