% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_name_to_abb.R
\name{US_name_to_abb}
\alias{US_name_to_abb}
\title{US State Abbreviation to Name}
\usage{
US_name_to_abb(x)
}
\arguments{
\item{x}{A character element or vector.}
}
\value{
A character element or vector.
}
\description{
This function converts a US state name to the two-letter state abbreviation.
It uses the base R state.abb and state.name vectors, but is extended to included
the District of Columbia.
}
\examples{
US_name_to_abb(c("Texas", "District of Columbia"))
}
\seealso{
\code{\link[=US_abb_to_name]{US_abb_to_name()}}
}
