% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_blog_template.R
\name{use_blog_template}
\alias{use_blog_template}
\title{Create folder and files for website blog}
\usage{
use_blog_template(date_chr = "2023-08-01", slug = "blog-title")
}
\arguments{
\item{date_chr}{Date of blog in yyyy-mm-dd format.}

\item{slug}{Character string for url.}
}
\value{
a message if template file was successfully copied over
}
\description{
Create folder and files for website blog
}
