% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tt_template.R
\name{use_tt_template}
\alias{use_tt_template}
\title{Create folder and files for #TidyTuesday data viz}
\usage{
use_tt_template(date_chr = "2023-08-01", lang = "R", readme = TRUE)
}
\arguments{
\item{date_chr}{Date of #TidyTuesday in yyyy-mm-dd format.}

\item{lang}{Language, either \code{"R"}, \code{"Python"}, or \code{"D3"}.}

\item{readme}{Boolean indicating whether to also use README template.}
}
\value{
a message if template file was successfully copied over
}
\description{
Create folder and files for #TidyTuesday data viz
}
