% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_missing.R
\name{viz_missing}
\alias{viz_missing}
\title{Visualise missing data}
\usage{
viz_missing(data, missing = NA, show_rownames = TRUE, show_colnames = TRUE)
}
\arguments{
\item{data}{Data to visualise. Needs to be a data frame or tibble.}

\item{missing}{List of values to define as missing. Default \code{NA}.}

\item{show_rownames}{Boolean to determine if row names are printed.
Default \code{TRUE}.}

\item{show_colnames}{Boolean to determine if column names are printed.
Default \code{TRUE}.}
}
\description{
Create a heatmap showing which values in a dataframe
are missing.
}
\examples{
viz_missing(mtcars)
}
