% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forest_plot}
\alias{forest_plot}
\title{Generate a forest plot}
\usage{
forest_plot(
  dat,
  col_num,
  group = NA,
  y_axis,
  units = NULL,
  title = NULL,
  scale = 1,
  null_at = 1,
  text_size = "norm",
  meta = FALSE,
  f_var = "facet_var"
)
}
\arguments{
\item{dat}{a data.frame containing the data}

\item{col_num}{}
}
\description{
Generate a forest plot using ggplot2. This function has capability to be split into multiple forests and was generated to display Mendelian randomization results with a binary outcome.
}
