% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tt_template.R
\name{use_tt_template}
\alias{use_tt_template}
\title{Create folder and files for #TidyTuesday data viz}
\usage{
use_tt_template(date_chr = "2023-08-01", readme = TRUE, lang = "R")
}
\arguments{
\item{date_chr}{Date of #TidyTuesday in yyyy-mm-dd format.}

\item{readme}{Boolean indicating whether to also use README template.}

\item{lang}{Language, either \code{"R"} or \code{"Python"}}
}
\value{
a message if template file was successfully copied over
}
\description{
Create folder and files for #TidyTuesday data viz
}
