% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{write_grm}
\alias{write_grm}
\title{Write readGRM style output back to binary GRM for use with GCTA}
\usage{
write_grm(grm, rootname, n.file = FALSE)
}
\arguments{
\item{grm}{Output from \link{readGRM}}

\item{n.file}{Write out grm.N.bin file - not always needed}
}
\description{
Write readGRM style output back to binary GRM for use with GCTA
}
